complete TeX.pool reorganization
