+++ rtx_package/src/engine/latex_ch2_document.rs
@@ -67,7 +67,7 @@ LoadDefinitions!({
         boxes.push(stomach::digest(ops)?);
       }
       // Should we try to indent the last paragraph? If so, it goes like this:
-      boxes.push(stomach::digest(T_CS!("\\normal@par"))?);
+      boxes.push(stomach::digest(T_CS!("\\lx@normal@par"))?);
       // Now we check whether we're down to the last stack frame.
       // It is common for unclosed { or even environments
       // and we want to at least compress & avoid unnecessary errors & warnings.
