+++ rtx_package/src/engine/tex.rs
@@ -3,8 +3,7 @@ LoadDefinitions!({
   // port of TeX.pool.ltxml
   // commit 4cd73e7584c5f0422293ba38f9b757332584afec
   // Author: Bruce Miller <nebconinc@gmail.com>
-  // Date:   Thu May 9 13:19:32 2024 -0400
-  
+  // Date:   Thu May 9 13:19:32 2024 -0400  
   InnerPool!(base_schema);
   InnerPool!(base_parameter_types);
   InnerPool!(base_utilities);
@@ -22,75 +21,18 @@ LoadDefinitions!({
   InnerPool!(tex_logic);
   InnerPool!(tex_macro);
   InnerPool!(tex_marks);
-  // -- CONTINUE HERE:
   InnerPool!(tex_math);
   InnerPool!(tex_scripts);
-  // InnerPool!(tex_page);
-  // InnerPool!(tex_paragraph);
-  // InnerPool!(tex_penalties);
-  // InnerPool!(tex_registers);
-  // InnerPool!(tex_tables);
-
+  InnerPool!(tex_page);
+  InnerPool!(tex_paragraph);
+  InnerPool!(tex_penalties);
   InnerPool!(tex_registers);
-  // lines 1217-1524
-  InnerPool!(tex_assignment);
-  // lines 2417-2785
-  InnerPool!(tex_ch24_primitives);
-  // lines 2786-3523
-  InnerPool!(tex_alignment);
-  // lines 3524-3660
-  InnerPool!(tex_para);
-  // lines 3661-3783
-  InnerPool!(tex_ch25_primitives);
-  // lines 4511-4688
-  InnerPool!(tex_math_style);
-  // lines 4689-5041
-  InnerPool!(tex_appendix_b_to_p349);
-  // lines 5042-5290
-  InnerPool!(tex_appendix_b_p350_to_p355);
-  // lines 5621-5655
-  InnerPool!(tex_appendix_b_p356);
-  // lines 5656-5783
-  InnerPool!(tex_accents);
-  // lines 5784-5832
-  InnerPool!(tex_appendix_b_p357);
-  // lines 5833-6278
-  InnerPool!(tex_appendix_b_p358);
-  // lines 6279-6329
-  InnerPool!(tex_appendix_b_p359);
-  // lines 6330-6377
-  InnerPool!(tex_math_accents);
-  // lines 6378-6574
-  InnerPool!(latex_delimiters);
-  // lines 6575-6629
-  InnerPool!(tex_appendix_b_p360);
-  // lines 6630-6714
-  InnerPool!(tex_appendix_b_p361);
-  // lines 6715-6960
-  InnerPool!(tex_appendix_b_p362);
-  // lines 6961-6998
-  InnerPool!(tex_appendix_b_p363);
-  // lines 6999-7010
-  InnerPool!(tex_appendix_b_p364);
+  InnerPool!(tex_tables);
+  InnerPool!(etex); // unless... ?
+  InnerPool!(pdftex); // unless... ?
 
-  //======================================================================
-  // End of TeX Book definitions.
-  //======================================================================
-
-  //**********************************************************************
-  // Stray stuff .... where to ?
-  //**********************************************************************
-
-  // lines 7013-7036
-  InnerPool!(tex_stray_math_style);
-  // lines 7037-7140
-  InnerPool!(tex_special_chars);
-  // lines 7721 - 7725
-  InnerPool!(etex);
-  InnerPool!(pdftex);
-
-  // should we port the deprecations to rust? postpone for now.
-  //InnerPool!(base_deprecated);
+  // TODO: should we port the deprecations to rust? postpone for now.
+  // InnerPool!(base_deprecated);
   InnerPool!(plain);
 
   //%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
@@ -103,9 +45,13 @@ LoadDefinitions!({
   Let!("\\protect", "\\relax");
   DefRegister!("\\everyhelp", Tokens!());
   DefMacro!("\\hiderel{}", "#1");    // Just ignore, for now...
-  // TODO: revisit organization for hooks
+  
   InnerPool!(latex_hook);
   
+  //======================================================================
+  // After all other rewrites have acted, a little cleanup
+  // [This suggests that it should be (one of) the LAST (math) rewrite applied?
+  // Do we need to define it last?]
   // DefRewrite(xpath => 'descendant-or-self::ltx:XMWrap[count(child::*)=1]',
   //   replace => sub { my ($document, $wrap) = @_;
   //     if (my $node = $document->getFirstChildElement($wrap)) {
