--- rtx_package/src/engine/tex_appendix_b_p364.rs
@@ -1,8 +1,0 @@
-use crate::prelude::*;
-
-//======================================================================
-// TeX Book, Appendix B. p. 364
-LoadDefinitions!({
-  // Let's hope nobody is messing with the output routine...
-  DefPrimitive!("\\footnoterule", None);
-});
