--- rtx_package/src/engine/tex_stray_math_style.rs
@@ -1,12 +1,0 @@
-use crate::prelude::*;
-
-LoadDefinitions!({
-  // Mostly ignorable, although it could add an attribute to an ancestor
-  // to record the desired justification.
-  // Spacing stuff
-  DefConstructor!("\\@", "");
-  // Math spacing.
-
-  // Math style.
-
-});
